First version of PAS select
