+++ src/components/QuestionnaireForm/QuestionnaireForm.jsx
@@ -10,10 +10,6 @@ import BooleanInput from "../Inputs/BooleanInput/BooleanInput";
 import QuantityInput from "../Inputs/QuantityInput/QuantityInput";
 import { findValueByPrefix } from "../../util/util.js";
 import OpenChoice from "../Inputs/OpenChoiceInput/OpenChoice";
-import SendDMEOrder from "../../util/DMEOrders";
-
-// Note: code to enable/disable DME Orders
-var dMEOrdersEnabled = false;
 
 export default class QuestionnaireForm extends Component {
   constructor(props) {
@@ -824,56 +820,55 @@ export default class QuestionnaireForm extends Component {
     console.log(priorAuthBundle);
 
     this.props.setPriorAuthClaim(priorAuthBundle);
-
-    const Http = new XMLHttpRequest();
-    const priorAuthUrl =
-      "https://davinci-prior-auth.logicahealth.org/fhir/Claim/$submit";
-    // const priorAuthUrl = "http://localhost:9000/fhir/Claim/$submit";
-    Http.open("POST", priorAuthUrl);
-    Http.setRequestHeader("Content-Type", "application/fhir+json");
-    Http.send(JSON.stringify(priorAuthBundle));
-    var qForm = this;
-    Http.onreadystatechange = function() {
-      if (this.readyState === XMLHttpRequest.DONE) {
-        var message =
-          "Prior Authorization Failed.\nNo ClaimResponse found within bundle.";
-        if (this.status === 201) {
-          var claimResponseBundle = JSON.parse(this.responseText);
-          var claimResponse = claimResponseBundle.entry[0].resource;
-          message = "Prior Authorization " + claimResponse.disposition + "\n";
-          message += "Prior Authorization Number: " + claimResponse.preAuthRef;
-
-          // DME Orders
-          if (dMEOrdersEnabled) SendDMEOrder(qForm, response);
-        } else {
-          console.log(this.responseText);
-          message = "Prior Authorization Request Failed.";
-        }
-        console.log(message);
-
-        // TODO pass the message to the PriorAuth page instead of having it query again
-        var patientEntry = claimResponseBundle.entry.find(function(entry) {
-          return entry.resource.resourceType == "Patient";
-        });
-
-        // fall back to resource.id if resource.identifier is not populated
-        var patientId;
-        if (patientEntry.resource.identifier == null) {
-          patientId = patientEntry.resource.id;
-        } else {
-          patientId = patientEntry.resource.identifier[0].value;
-        }
-        let priorAuthUri =
-          "priorauth?identifier=" +
-          claimResponse.preAuthRef +
-          "&patient.identifier=" +
-          patientId;
-        console.log(priorAuthUri);
-        window.location.href = priorAuthUri;
-      }
-    };
-
     localStorage.removeItem(response.questionnaire);
+
+    // const Http = new XMLHttpRequest();
+    // const priorAuthUrl =
+    //   "https://davinci-prior-auth.logicahealth.org/fhir/Claim/$submit";
+    // // const priorAuthUrl = "http://localhost:9000/fhir/Claim/$submit";
+    // Http.open("POST", priorAuthUrl);
+    // Http.setRequestHeader("Content-Type", "application/fhir+json");
+    // Http.send(JSON.stringify(priorAuthBundle));
+    // var qForm = this;
+    // Http.onreadystatechange = function() {
+    //   if (this.readyState === XMLHttpRequest.DONE) {
+    //     var message =
+    //       "Prior Authorization Failed.\nNo ClaimResponse found within bundle.";
+    //     if (this.status === 201) {
+    //       var claimResponseBundle = JSON.parse(this.responseText);
+    //       var claimResponse = claimResponseBundle.entry[0].resource;
+    //       message = "Prior Authorization " + claimResponse.disposition + "\n";
+    //       message += "Prior Authorization Number: " + claimResponse.preAuthRef;
+
+    //       // DME Orders
+    //       if (dMEOrdersEnabled) SendDMEOrder(qForm, response);
+    //     } else {
+    //       console.log(this.responseText);
+    //       message = "Prior Authorization Request Failed.";
+    //     }
+    //     console.log(message);
+
+    //     // TODO pass the message to the PriorAuth page instead of having it query again
+    //     var patientEntry = claimResponseBundle.entry.find(function(entry) {
+    //       return entry.resource.resourceType == "Patient";
+    //     });
+
+    //     // fall back to resource.id if resource.identifier is not populated
+    //     var patientId;
+    //     if (patientEntry.resource.identifier == null) {
+    //       patientId = patientEntry.resource.id;
+    //     } else {
+    //       patientId = patientEntry.resource.identifier[0].value;
+    //     }
+    //     let priorAuthUri =
+    //       "priorauth?identifier=" +
+    //       claimResponse.preAuthRef +
+    //       "&patient.identifier=" +
+    //       patientId;
+    //     console.log(priorAuthUri);
+    //     window.location.href = priorAuthUri;
+    //   }
+    // };
   }
 
   isEmptyAnswer(answer) {
